+++ src/tools/releases.rs
@@ -1,5 +1,6 @@
 //! Release tool request types
 
+use crate::serde::flexible_u64_opt;
 use rmcp::schemars;
 use schemars::JsonSchema;
 use serde::Deserialize;
@@ -21,7 +22,8 @@ pub struct ListReleasesRequest {
 
     /// Maximum number of releases to return
     #[schemars(description = "Maximum number of releases to return (default: 30)")]
-    pub limit: Option<u32>,
+    #[serde(default, deserialize_with = "flexible_u64_opt")]
+    pub limit: Option<u64>,
 }
 
 /// Request parameters for get_release tool
